Move environment variable class out of beans directory.
